requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==3.1.4
marshmallow==3.20.1
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
WTForms==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
